컴포넌트 분리 & scss 분리
