+++ src/Components/Detail/Detail/Tab.scss
@@ -1,0 +1,60 @@
+@import "../../../Styles/common";
+
+.Detail {
+  .tab {
+    width: 1010px;
+    margin-top: 40px;
+    ul {
+      display: flex;
+      flex-direction: row;
+      width: 100%;
+      z-index: 1;
+      .tabOn {
+        box-sizing: border-box;
+        background-color: white;
+        border-left: 1px solid #bfbfbf;
+        border-top: 1px solid #bfbfbf;
+        border-right: 1px solid #f0eef1;
+        width: 170px;
+        height: 46px;
+        text-align: center;
+        line-height: 46px;
+        color: #4c4c4c;
+        font-size: 14px;
+        z-index: 10;
+        cursor: pointer;
+      }
+      .tabOff {
+        box-sizing: border-box;
+        border-left: 1px solid #bfbfbf;
+        border-right: 1px solid #f0eef1;
+        border-top: 1px solid #bfbfbf;
+        border-bottom: 1px solid #bfbfbf;
+        background-color: #f0eef1;
+        width: 170px;
+        height: 46px;
+        text-align: center;
+        line-height: 46px;
+        color: #4c4c4c;
+        font-size: 14px;
+        cursor: pointer;
+      }
+      .tabLast {
+        box-sizing: border-box;
+        border: 1px solid #bfbfbf;
+        background-color: #f0eef1;
+        width: 170px;
+        height: 46px;
+        text-align: center;
+        line-height: 46px;
+        color: #4c4c4c;
+        font-size: 14px;
+        cursor: pointer;
+      }
+      .line {
+        width: 180px;
+        border-bottom: 1px solid #bfbfbf;
+      }
+    }
+  }
+}
