+++ src/Components/Detail/Detail/WhyKurly.scss
@@ -1,0 +1,81 @@
+@import "../../../Styles/common";
+
+.Detail {
+  .why-kurly {
+    width: 1010px;
+    text-align: center;
+    table {
+      background: #e0e0e0;
+      margin: 70px 0;
+      width: 100%;
+      text-align: left;
+      font-family: "Dotum";
+      font-weight: normal;
+      word-spacing: -3px;
+      border-collapse: collapse;
+      border-spacing: 0;
+      tr {
+        th {
+          width: 15%;
+          background: #f5f5f5;
+          color: #515151;
+          font-weight: normal;
+          padding-left: 15px;
+          font-size: 11px;
+          height: 28px;
+          padding: 5px 0 5px 10px;
+          border-width: 1px;
+          line-height: 20px;
+        }
+        td {
+          width: 35%;
+          color: #666666;
+          padding-left: 15px;
+          background: #ffffff;
+          font-size: 11px;
+          height: 28px;
+          padding: 5px 0 5px 10px;
+          border-width: 1px;
+          line-height: 20px;
+        }
+      }
+    }
+    h3 {
+      font-size: 33px;
+      font-weight: 400;
+      color: #666;
+      margin: 70px 0 90px 0;
+    }
+    .why {
+      display: flex;
+      flex-direction: row;
+      align-items: flex-start;
+      justify-content: space-around;
+      margin-bottom: 70px;
+    }
+    .why-kurly-contents {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      .why-kurly-info {
+        padding-bottom: 20px;
+        margin: 0 30px;
+      }
+      .title {
+        color: $color;
+        font-weight: 500;
+        display: block;
+        font-size: 18px;
+        text-align: center;
+        margin-top: 10px;
+      }
+      .content {
+        color: #333;
+        padding: 0;
+        font-size: 16px;
+        font-weight: 400;
+        line-height: 25px;
+      }
+    }
+  }
+}
