+++ src/Components/Detail/ProductInfo/Count.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./Count.scss";
 
 export default class Count extends React.Component {
   render() {
