+++ src/Components/Detail/ProductInfo/Count.scss
@@ -1,0 +1,34 @@
+@import "../../../Styles/common";
+
+.Detail {
+  .option {
+    display: flex;
+    flex-direction: row;
+    width: 86px;
+    border: 1px solid #dddfe1;
+    border-radius: 3px;
+    button {
+      margin: 0;
+      padding: 0;
+      width: 22px;
+      height: 22px;
+      background: #f6f7f7;
+      color: $color;
+      outline: none;
+      border: 0;
+      cursor: pointer;
+    }
+
+    .leftBtn {
+      border-right: 1px solid #dddfe1;
+      font-size: 24px;
+      line-height: 0px;
+    }
+
+    .rightBtn {
+      border-left: 1px solid #dddfe1;
+      font-size: 20px;
+      line-height: 0;
+    }
+  }
+}
