+++ src/Components/Detail/ReviewQA/Review.js
@@ -1,14 +1,15 @@
 import React, { Component } from "react";
-import Table from "../Table";
+import Table from "./Table";
 import PageBtn from "./PageBtn";
+import "./ReviewQa.scss";
 
-export default class Qa extends Component {
+export default class Review extends Component {
   render() {
     return (
-      <div className="qa">
+      <div className="review">
         <div className="product-review">
           <div className="">
-            <h2>PRODUCT Q&A</h2>
+            <h2>PRODUCT REVIEW</h2>
             <ul>
               <li>
                 <div className="ico" />
@@ -36,13 +37,8 @@ export default class Qa extends Component {
         </div>
 
         <Table />
-        <div className="qa-btn">
-          <div>
-            <button className="write-all-view">전체보기</button>
-          </div>
-          <div>
-            <button className="write-qa">상품문의</button>
-          </div>
+        <div className="write">
+          <button className={"write-btn"}>후기쓰기</button>
         </div>
         <PageBtn />
       </div>
