+++ src/Components/Detail/Slide/DetailSlide.js
