+++ src/Components/Popup/PopUp.scss
@@ -1,0 +1,68 @@
+@import "../../Styles/common";
+
+.Detail {
+  .PopUp {
+    position: fixed;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    margin: auto;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 1000000;
+    .msg {
+      position: absolute;
+      width: 490px;
+      background-color: #fff;
+      text-align: center;
+      top: 30%;
+      left: 33%;
+      .msg-box {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: space-between;
+        background-color: #fff;
+        margin: 20px 25px 0 25px;
+        padding-bottom: 20px;
+        border-bottom: 1px solid #b2b2b2;
+        h3 {
+          font-size: 14px;
+          font-weight: 500;
+          color: $color;
+        }
+        .close {
+          width: 19px;
+          height: 19px;
+          cursor: pointer;
+          background-image: url(https://res.kurly.com/pc/etc/old/images/common/icon-close-button.png);
+        }
+      }
+      .contents {
+        font-size: 13px;
+        margin: 50px 0;
+        color: #4c4c4c;
+      }
+      .btn {
+        height: 105px;
+        background-color: #f5f5f5;
+        padding: 30px 0;
+        box-sizing: border-box;
+        display: flex;
+        justify-content: center;
+        button {
+          font-size: 13px;
+          width: 150px;
+          height: 45px;
+          outline: none;
+          background-color: $color;
+          border: 1px solid $color;
+          color: #fff;
+          cursor: pointer;
+        }
+      }
+    }
+  }
+}
